+++ .gitignore
@@ -1,4 +1,5 @@
 out
+runinfo
 .ipynb_checkpoints
 __pycache__
 notes*
