--- README.md
@@ -1,3 +1,0 @@
-# IFU-M
-
-readme coming soon (:
