+++ ifum_pkg/ifum.egg-info/SOURCES.txt
@@ -1,0 +1,17 @@
+README.md
+pyproject.toml
+ifum/__init__.py
+ifum/calibrate.py
+ifum/config.py
+ifum/helper.py
+ifum/hexagonify.py
+ifum/maskopt.py
+ifum/rectify.py
+ifum/stack.py
+ifum/stitch.py
+ifum/utils.py
+ifum.egg-info/PKG-INFO
+ifum.egg-info/SOURCES.txt
+ifum.egg-info/dependency_links.txt
+ifum.egg-info/requires.txt
+ifum.egg-info/top_level.txt
