+++ ifum_pkg/ifum.egg-info/requires.txt
@@ -1,0 +1,9 @@
+notebook
+pandas
+numpy
+scikit-learn
+matplotlib
+jupyter
+parsl
+astropy
+scipy
