+++ ifum_pkg/ifum/ifum.egg-info/SOURCES.txt
@@ -1,0 +1,9 @@
+README.md
+pyproject.toml
+src/__init__.py
+src/pipeline_helper.py
+src/ifum.egg-info/PKG-INFO
+src/ifum.egg-info/SOURCES.txt
+src/ifum.egg-info/dependency_links.txt
+src/ifum.egg-info/requires.txt
+src/ifum.egg-info/top_level.txt
