+++ ifum_pkg/ifum/ifum.egg-info/requires.txt
@@ -1,0 +1,6 @@
+notebook
+pandas
+numpy
+scikit-learn
+matplotlib
+jupyter
