+++ ifum_pkg/pyproject.toml
@@ -1,0 +1,28 @@
+[build-system]
+requires = ["setuptools >= 77.0.3"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ifum"
+version = "2025.0.0"
+description = "Parsl integration in an integral field unit spectroscopy pipeline."
+authors = [
+    {name = "Daniel Babnigg"}
+]
+readme = "README.md"
+requires-python = ">=3.7"
+
+dependencies = [
+    "notebook",
+    "pandas",
+    "numpy",
+    "scikit-learn",
+    "matplotlib",
+    "jupyter",
+    "parsl",
+    "astropy",
+    "scipy"
+]
+
+[tool.jupyter]
+notebook = true
