+++ parsl_py38.yml
@@ -1,0 +1,170 @@
+name: parsl_py38
+channels:
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - asttokens=3.0.0=py39h06a4308_0
+  - backcall=0.2.0=pyhd3eb1b0_0
+  - ca-certificates=2025.2.25=h06a4308_0
+  - comm=0.2.1=py39h06a4308_0
+  - debugpy=1.5.1=py39h295c915_0
+  - decorator=5.1.1=pyhd3eb1b0_0
+  - entrypoints=0.4=py39h06a4308_0
+  - exceptiongroup=1.2.0=py39h06a4308_0
+  - executing=0.8.3=pyhd3eb1b0_0
+  - ipykernel=6.19.2=py39hb070fc8_0
+  - ipython=8.15.0=py39h06a4308_0
+  - jedi=0.19.2=py39h06a4308_0
+  - jupyter_core=5.7.2=py39h06a4308_0
+  - ld_impl_linux-64=2.40=h12ee557_0
+  - libffi=3.3=he6710b0_2
+  - libgcc-ng=9.1.0=hdf63c60_0
+  - libsodium=1.0.18=h7b6447c_0
+  - libstdcxx-ng=9.1.0=hdf63c60_0
+  - matplotlib-inline=0.1.6=py39h06a4308_0
+  - ncurses=6.3=h7f8727e_2
+  - nest-asyncio=1.6.0=py39h06a4308_0
+  - openssl=1.1.1w=h7f8727e_0
+  - packaging=24.2=py39h06a4308_0
+  - parso=0.8.4=py39h06a4308_0
+  - pexpect=4.9.0=py39h06a4308_0
+  - pickleshare=0.7.5=pyhd3eb1b0_1003
+  - pip=25.1=pyhc872135_2
+  - platformdirs=4.3.7=py39h06a4308_0
+  - prompt-toolkit=3.0.43=py39h06a4308_0
+  - ptyprocess=0.7.0=pyhd3eb1b0_2
+  - pure_eval=0.2.2=pyhd3eb1b0_0
+  - pygments=2.19.1=py39h06a4308_0
+  - python=3.9.12=h12debd9_1
+  - python-dateutil=2.9.0post0=py39h06a4308_2
+  - readline=8.1.2=h7f8727e_1
+  - setuptools=78.1.1=py39h06a4308_0
+  - six=1.17.0=py39h06a4308_0
+  - sqlite=3.38.5=hc218d9a_0
+  - stack_data=0.2.0=pyhd3eb1b0_0
+  - tk=8.6.12=h1ccaba5_0
+  - traitlets=5.14.3=py39h06a4308_0
+  - wcwidth=0.2.5=pyhd3eb1b0_0
+  - wheel=0.45.1=py39h06a4308_0
+  - xz=5.2.5=h7f8727e_1
+  - zeromq=4.3.4=h2531618_0
+  - zlib=1.2.12=h7f8727e_2
+  - pip:
+    - anyio==4.9.0
+    - argon2-cffi==25.1.0
+    - argon2-cffi-bindings==21.2.0
+    - arrow==1.3.0
+    - astropy==6.0.1
+    - astropy-iers-data==0.2025.6.23.0.39.50
+    - async-lru==2.0.5
+    - attrs==25.3.0
+    - babel==2.17.0
+    - beautifulsoup4==4.13.4
+    - bleach==6.2.0
+    - blinker==1.9.0
+    - certifi==2025.6.15
+    - cffi==1.17.1
+    - charset-normalizer==3.4.2
+    - click==8.1.8
+    - contourpy==1.3.0
+    - cryptography==45.0.4
+    - cycler==0.12.1
+    - defusedxml==0.7.1
+    - dill==0.4.0
+    - fastjsonschema==2.21.1
+    - filelock==3.18.0
+    - flask==3.1.1
+    - flask-sqlalchemy==3.1.1
+    - fonttools==4.58.5
+    - fqdn==1.5.1
+    - globus-sdk==3.58.0
+    - greenlet==3.2.3
+    - h11==0.16.0
+    - httpcore==1.0.9
+    - httpx==0.28.1
+    - idna==3.10
+    - ifum==2025.0.0
+    - importlib-metadata==8.7.0
+    - importlib-resources==6.5.2
+    - ipywidgets==8.1.7
+    - isoduration==20.11.0
+    - itsdangerous==2.2.0
+    - jinja2==3.1.6
+    - joblib==1.5.1
+    - json5==0.12.0
+    - jsonpointer==3.0.0
+    - jsonschema==4.24.0
+    - jsonschema-specifications==2025.4.1
+    - jupyter==1.1.1
+    - jupyter-client==8.6.3
+    - jupyter-console==6.6.3
+    - jupyter-events==0.12.0
+    - jupyter-lsp==2.2.5
+    - jupyter-server==2.16.0
+    - jupyter-server-terminals==0.5.3
+    - jupyterlab==4.4.4
+    - jupyterlab-pygments==0.3.0
+    - jupyterlab-server==2.27.3
+    - jupyterlab-widgets==3.0.15
+    - kiwisolver==1.4.7
+    - markupsafe==3.0.2
+    - matplotlib==3.9.4
+    - mistune==3.1.3
+    - narwhals==1.44.0
+    - nbclient==0.10.2
+    - nbconvert==7.16.6
+    - nbformat==5.10.4
+    - networkx==3.2.1
+    - notebook==7.4.4
+    - notebook-shim==0.2.4
+    - numpy==1.26.4
+    - overrides==7.7.0
+    - pandas==2.3.0
+    - pandocfilters==1.5.1
+    - parsl==2025.6.16
+    - pillow==11.3.0
+    - plotly==6.1.2
+    - prometheus-client==0.22.1
+    - psutil==7.0.0
+    - pycparser==2.22
+    - pyerfa==2.0.1.5
+    - pyjwt==2.10.1
+    - pyparsing==3.2.3
+    - python-json-logger==3.3.0
+    - pytz==2025.2
+    - pyyaml==6.0.2
+    - pyzmq==27.0.0
+    - referencing==0.36.2
+    - requests==2.32.4
+    - rfc3339-validator==0.1.4
+    - rfc3986-validator==0.1.1
+    - rpds-py==0.26.0
+    - scikit-learn==1.6.1
+    - scipy==1.13.1
+    - send2trash==1.8.3
+    - setproctitle==1.3.6
+    - sniffio==1.3.1
+    - sortedcontainers==2.4.0
+    - soupsieve==2.7
+    - sqlalchemy==2.0.41
+    - tblib==3.1.0
+    - terminado==0.18.1
+    - threadpoolctl==3.6.0
+    - tinycss2==1.4.0
+    - tomli==2.2.1
+    - tornado==6.5.1
+    - tqdm==4.67.1
+    - typeguard==4.4.3
+    - types-python-dateutil==2.9.0.20250708
+    - typing-extensions==4.14.1
+    - tzdata==2025.2
+    - uri-template==1.3.0
+    - urllib3==2.4.0
+    - webcolors==24.11.1
+    - webencodings==0.5.1
+    - websocket-client==1.8.0
+    - werkzeug==3.1.3
+    - widgetsnbextension==4.0.14
+    - zipp==3.23.0
+prefix: /home/babnigg/.conda/envs/parsl_py38
+
